refactor: refactor the image crawling process
