+++ FamilyStoryADA/Data/SampleImageCrawl/Datasource/DTO/DeleteResponseObject.swift
@@ -1,0 +1,11 @@
+//
+//  DeleteResponseObject.swift
+//  FamilyStoryADA
+//
+//  Created by Daniel Ian Kurniawan on 03/10/24.
+//
+
+struct DeleteResponseObject: Codable {
+    let status: String
+    let message: String
+}
