+++ FamilyStoryADA/FamilyStoryADAApp.swift
@@ -11,8 +11,8 @@ import SwiftUI
 struct FamilyStoryADAApp: App {
     var body: some Scene {
         WindowGroup {
-//            ExampleView()
-            CameraView()
+            SampleImageCrawlView()
+//            CameraView()
         }
     }
 }
